Board header separation
